% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCCESS.R
\name{encode}
\alias{encode}
\title{Generates an encoded subspace of a single-cell RNA-seq expression matrix}
\usage{
encode(
  dat,
  seed = 1,
  max_random_projection = 2048,
  encoded_dim = 16,
  hidden_dims = c(128),
  learning_rate = 0.001,
  batch_size = 32,
  epochs = 100,
  verbose = 2,
  scale = FALSE,
  genes_as_rows = FALSE,
  cores = 1
)
}
\arguments{
\item{dat}{A raw data matrix, data frame or tibble containing scRNA-seq expression values. By default,
genes are assumed to be represented by columns and samples are assumed to be
represented by rows (but see the argument genes_as_rows). NA values are not
supported, but may be replaced by 0s. It could be a filtered matrix by prefilter() funtion.}

\item{seed}{Random seed for initial gene sampling. Currently a seed cannot
be set to reproducibly determine the behaviour of the autoencoder artificial
neural network.}

\item{max_random_projection}{Determines the maximum number of genes to be initially sampled prior to
autoencoder training. In practice the number of genes sampled is equal to this
number or 80% of the genes present in the matrix (rounded up), whichever is
smaller.}

\item{encoded_dim}{The number of features in the encoded data.}

\item{hidden_dims}{A vector of 1 or more integers, representing the number of nodes in each
successive hidden layer of the encoder half of the autoencoder. Hidden layers in
the decoder use these widths in reverse.}

\item{learning_rate}{Learning rate for training the artificial neural network.}

\item{batch_size}{Number of samples per training batch.}

\item{epochs}{Number of training epochs.}

\item{verbose}{Determines the verbosity of the keras training function, 0 is silent, 1 is an animated progress bar, 2 is message, default is 2}

\item{scale}{If TRUE, gene values are rescaled to a mean of 0 and a standard deviation of 1.}

\item{genes_as_rows}{If TRUE, rows in the expression matrix are assumed to represent genes and columns
assumed to represent cells.}

\item{cores}{number of cores}
}
\value{
An encoded expression matrix wherein cells are represented by rows and latent features are represented by columns.
}
\description{
This function accepts a single scRNA-seq expression matrix,
randomly samples a number of genes without replacement
and trains an autoencoder artificial neural network on the resulting data.
The function uses part of this network to encode cell data within
a lower-dimensional latent space and returns the encoded matrix.
This function does not need to be called directly by the user for clustering (see ensemble_cluster function below),
but is provided for greater flexibility.
It is not recommended to run this function in parallel
as model training makes use of resources in parallel (CPU cores or GPU, depending on computer setup).
}
\examples{

# library(SingleCellExperiment)
# library(scCCESS)
# data("sce", package = "scCCESS")
# dat=SingleCellExperiment::counts(sce)
# dat.filtered=prefilter(dat)
# encoding = encode(dat.filtered, seed = 1,
#                   genes_as_rows = T, ensemble_sizes = 10,
#                   verbose = 0, scale = F, batch_size = 64,
#                   cores=8)

}
