% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCCESS.R
\name{estimate_k}
\alias{estimate_k}
\title{Estimated number of clusters via clustering stability metric}
\usage{
estimate_k(
  dat,
  seed = 1,
  criteria_method = "NMI",
  cluster_func = function(x) kmeans(x, centers = 5),
  krange = 2:25,
  ensemble_sizes = c(1, 5, 10, 20, 50),
  cores = 1,
  ...
)
}
\arguments{
\item{dat}{input scRNAseq count data}

\item{seed}{radom seed}

\item{criteria_method}{criteria used to measure stability}

\item{cluster_func}{clustering function}

\item{krange}{range of k}

\item{ensemble_sizes}{ensemble size}

\item{cores}{number of cores}

\item{...}{other parameters that used in encode function}
}
\value{
the number of cell types
}
\description{
Estimated number of clusters via clustering stability metric
}
