% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCCESS.R
\name{prefilter}
\alias{prefilter}
\title{pre-filtering function
Filter input dataset}
\usage{
prefilter(
  table,
  minReads = 1000,
  minGene = 100,
  minCountsperGene = 1,
  removeZeroGene = T
)
}
\arguments{
\item{table}{input scRNAseq matrix, where each row represents a gene
and each column represents a single cell with a raw count for every row (gene) in the matrix.}

\item{minReads}{minimum reads count per cell}

\item{minGene}{minimum gene that have counts more than \code{minCountsperGene}}

\item{minCountsperGene}{at least how many counts are required for each gene}

\item{removeZeroGene}{remove genes that has zero counts were detected in the cell}
}
\value{
a filtered scRNAseq expression matrix
}
\description{
pre-filtering function
Filter input dataset
}
