% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{sce: a random sampled subset of Tabula Muris}
\format{
An SingleCellExperiment object of scRNAseq, includes 10 cell types and 1000 cells.
}
\source{
\url{https://figshare.com/projects/Tabula_Muris_Transcriptomic_characterization_of_20_organs_and_tissues_from_Mus_musculus_at_single_cell_resolution/27733}
}
\usage{
data(sce, package = 'scCCESS')
}
\description{
Data from Consortium, T.M., et al. Smart-seq2 Tabula Muris sampling dataset,
which is a SingleCellExperiment object include raw counts and cellTypes.
}
\references{
Consortium, T.M., et al.: Single-cell transcriptomics of 20 mouse organs creates a tabula muris.
Nature 562(7727), 367–372 (2018). doi:10.1038/s41586-018-0590-4
}
\keyword{datasets}
