% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCCESS.R
\name{ensemble_cluster}
\alias{ensemble_cluster}
\title{Generates an ensemble clustering of a single-cell RNA-seq expression matrix}
\usage{
ensemble_cluster(
  dat,
  seed = 1,
  cluster_func = function(x) kmeans(x, centers = 5),
  ensemble_sizes = c(1, 5, 10, 20, 50),
  cores = 1,
  ...
)
}
\arguments{
\item{dat}{a raw scRNAseq matrix, it could be a filtered matrix by  prefilter()  funtion}

\item{seed}{random seed}

\item{cluster_func}{clustering function, could be any function that need provide a range of k, such as K-means, SIMLR, C-means, K-medoids, etc.}

\item{ensemble_sizes}{ensemble size}

\item{cores}{number of cores}

\item{...}{other parameters that used in encode function}
}
\value{
ensemble cluster of all projections
}
\description{
Generates an ensemble clustering of a single-cell RNA-seq expression matrix
}
