% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCCESS.R
\name{prefilter}
\alias{prefilter}
\title{prefilter}
\usage{
prefilter(
  table,
  minReads = 1000,
  minGene = 100,
  minCountsperGene = 1,
  removeZeroGene = T
)
}
\arguments{
\item{table}{input scRNAseq matrix, where each row represents a gene
and each column represents a single cell with a raw count for every row (gene) in the matrix.}

\item{minReads}{minimum reads count (UMI) per cell}

\item{minGene}{minimum gene that have counts (UMI) more than \code{minCountsperGene}}

\item{minCountsperGene}{minimum counts (UMI) are required for each gene}

\item{removeZeroGene}{remove genes that has zero counts were detected in the cell}
}
\value{
a filtered scRNAseq expression matrix
}
\description{
Quality control: to filter the data to only include true cells that are of high quality
}
\examples{

# library(SingleCellExperiment)
# library(scCCESS)
# data("sce", package = "scCCESS")
# dat=SingleCellExperiment::counts(sce)
# dat.filtered=prefilter(dat)

}
